+++ __tests__/helpers/next-server.js
@@ -4,13 +4,25 @@
  */
 
 class ResponseCookies {
-  constructor() {
+  constructor(headers) {
     /** @type {Map<string, string>} */
     this._map = new Map()
+    /** @type {Headers} */
+    this._headers = headers
   }
 
-  set(name, value) {
+  set(name, value, options = {}) {
     this._map.set(name, value)
+
+    let cookieStr = `${name}=${encodeURIComponent(value)}`
+    if (options.httpOnly) cookieStr += '; HttpOnly'
+    if (options.secure) cookieStr += '; Secure'
+    if (options.sameSite) cookieStr += `; SameSite=${options.sameSite}`
+    if (options.path) cookieStr += `; Path=${options.path}`
+    if (options.domain) cookieStr += `; Domain=${options.domain}`
+    if (typeof options.maxAge === 'number') cookieStr += `; Max-Age=${options.maxAge}`
+
+    this._headers.append('Set-Cookie', cookieStr)
   }
 
   get(name) {
@@ -26,7 +38,7 @@ class ResponseCookies {
 export class NextResponse extends Response {
   constructor(body, init = {}) {
     super(body, init)
-    this.cookies = new ResponseCookies()
+    this.cookies = new ResponseCookies(this.headers)
   }
 
   static json(body, init = {}) {
@@ -45,11 +57,16 @@ export class NextResponse extends Response {
       headers: { Location: String(url) },
     })
   }
+
+  static next(init = {}) {
+    return new NextResponse(null, { status: init?.status ?? 200, headers: init?.headers })
+  }
 }
 
 export class NextRequest extends Request {
   constructor(input, init = {}) {
     super(input, init)
+    this.nextUrl = new URL(this.url)
     const cookieHeader = this.headers.get('cookie') || ''
     const jar = Object.fromEntries(
       cookieHeader
@@ -67,6 +84,7 @@ export class NextRequest extends Request {
         const value = jar[name]
         return value !== undefined ? { name, value } : undefined
       },
+      getAll: () => Object.entries(jar).map(([name, value]) => ({ name, value })),
     }
   }
 }